refactor: move renderer dirs up a level
