+++ src/picker/picker.html
@@ -4,13 +4,12 @@
 <head>
   <meta charset="UTF-8">
   <title>Browserosaurus</title>
-  <link rel="stylesheet" href="../../../build/main.css">
+  <link rel="stylesheet" href="../../build/main.css">
 
   <style>
     body {
       overflow: hidden;
       margin: 0;
-      /* background-color: #21252b; */
     }
 
   </style>
