+++ src/prefs/modules/UpdateStatus/UpdateStatus.js
