+++ src/prefs/prefs.html
@@ -4,7 +4,7 @@
 <head>
   <meta charset="UTF-8">
   <title>Browserosaurus Preferences</title>
-  <link rel="stylesheet" href="../../../build/main.css">
+  <link rel="stylesheet" href="../../build/main.css">
 
   <style>
     body {
@@ -13,7 +13,6 @@
       padding-right: 1rem;
       padding-bottom: 1rem;
       padding-left: 1rem;
-      background-color: #21252b;
     }
 
   </style>
