+++ src/utils/EscapeToHide.js
